requirements.txt
genson==1.2.2
graphviz==0.20.1
mitmproxy==10.2.4
mitmproxy-macos==0.5.1
beautifulsoup4~=4.12.3
html2image==2.0.4.3
dhash==1.4
pillow~=10.3.0
requests~=2.31.0
tabulate~=0.9.0
selenium==4.2.0
jsondiff==2.0.0
python-slugify==8.0.4